+++ client/cache.go
@@ -8,18 +8,18 @@ import (
 	"github.com/adrg/xdg"
 )
 
-func getCacheDirPath() string {
+func GetCacheDirPath() string {
 	dir := path.Join(xdg.CacheHome, "k8s-ldap-auth")
 	return dir
 }
 
-func getCacheFilePath() string {
-	file := path.Join(getCacheDirPath(), "token")
+func GetCacheFilePath() string {
+	file := path.Join(GetCacheDirPath(), "token")
 	return file
 }
 
 func getCachedToken() []byte {
-	token, err := ioutil.ReadFile(getCacheFilePath())
+	token, err := ioutil.ReadFile(GetCacheFilePath())
 	if err != nil {
 		return nil
 	}
@@ -28,11 +28,11 @@ func getCachedToken() []byte {
 }
 
 func refreshCache(data []byte) error {
-	if err := os.MkdirAll(getCacheDirPath(), 0750); err != nil {
+	if err := os.MkdirAll(GetCacheDirPath(), 0750); err != nil {
 		return err
 	}
 
-	if err := ioutil.WriteFile(getCacheFilePath(), data, 0640); err != nil {
+	if err := ioutil.WriteFile(GetCacheFilePath(), data, 0640); err != nil {
 		return err
 	}